added applicants validation on job edit

+++ app/Http/Controllers/API/Admin/JobController.php
@@ -267,6 +267,11 @@ class JobController extends Controller
         DB::beginTransaction();
         try
         { 
+            if(JobApplication::where('job_id', $job->id)->count()>0)
+            {
+                return response()->json(prepareResult(true, 'You cannot edit this job because this job have some applicants.', getLangByLabelGroups('messages','message_validation')), config('http_response.internal_server_error'));
+            }
+            
             //delete tags before update
             JobTag::where('job_id', $job->id)->delete();
